requirements.txt
#this includes prod and test dependencies 
# prior to shipping final prod seperate dev and prod dependencies 

annotated-types==0.7.0

anthropic==0.74.0

anyio==4.11.0
certifi==2025.11.12
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.0

fastapi==0.115.0

filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
packaging==25.0
pluggy==1.6.0

pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2

pytest==9.0.1
python-dotenv==1.0.1

PyYAML==6.0.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.6
tokenizers==0.22.1
tomli==2.3.0
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0

uvicorn==0.30.6

uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
